+++ docs/CHANGELOG.md
@@ -329,3 +329,4 @@ START OF THE PROJECT
 **v. x.x.x (xx.xx.xx)**
 - Allowed default value for Enum fields
 - Changed GenerateClass default value representation
+- Extended GenerateModel to generalize complex type build of different types (ex. null, string, record) to object